+++ src/main/rx/rx_express_lrs.c
@@ -1,9 +1,8 @@
-#include "rx.h"
+#include "rx_express_lrs.h"
 
 #include <string.h>
 
 #include "control.h"
-#include "drv_spi_sx12xx.h"
 #include "drv_time.h"
 #include "project.h"
 #include "usb_configurator.h"
@@ -26,57 +25,6 @@
 //  7     0     allocate the entire FIFO buffer for RX only
 #define FIFO_RX_BASE_ADDR_MAX 0b00000000
 
-#define NR_FHSS_ENTRIES (sizeof(fhss_freqs) / sizeof(uint32_t))
-#define NR_SEQUENCE_ENTRIES 256
-
-#define FREQ_STEP 61.03515625
-#define FHSS_RNG_MAX 0x7FFF
-
-typedef enum {
-  TLM_RATIO_NO_TLM = 0,
-  TLM_RATIO_1_2 = 2,
-  TLM_RATIO_1_4 = 4,
-  TLM_RATIO_1_8 = 8,
-  TLM_RATIO_1_16 = 16,
-  TLM_RATIO_1_32 = 32,
-  TLM_RATIO_1_64 = 64,
-  TLM_RATIO_1_128 = 128,
-} expresslrs_tlm_ratio_t;
-
-const expresslrs_tlm_ratio_t tlm_ration_map[] = {
-    TLM_RATIO_NO_TLM,
-    TLM_RATIO_1_2,
-    TLM_RATIO_1_4,
-    TLM_RATIO_1_8,
-    TLM_RATIO_1_16,
-    TLM_RATIO_1_32,
-    TLM_RATIO_1_64,
-    TLM_RATIO_1_128,
-};
-
-typedef enum {
-  RATE_500HZ = 0,
-  RATE_250HZ = 1,
-  RATE_200HZ = 2,
-  RATE_150HZ = 3,
-  RATE_100HZ = 4,
-  RATE_50HZ = 5,
-  RATE_25HZ = 6,
-  RATE_4HZ = 7
-} expresslrs_rf_rates_t;
-
-typedef struct {
-  int8_t index;
-  expresslrs_rf_rates_t rate; // Max value of 16 since only 4 bits have been assigned in the sync package.
-  sx12xx_bandwidth_t bw;
-  sx12xx_spreading_factor_t sf;
-  sx12xx_coding_rate_t cr;
-  uint32_t interval;                   //interval in us seconds that corresponds to that frequnecy
-  expresslrs_tlm_ratio_t tlm_interval; // every X packets is a response TLM packet, should be a power of 2
-  uint8_t fhss_hop_interval;           // every X packets we hope to a new frequnecy. Max value of 16 since only 4 bits have been assigned in the sync package.
-  uint8_t preamble_len;
-} expresslrs_mod_settings_t;
-
 static expresslrs_mod_settings_t air_rate_config[4] = {
     {0, RATE_200HZ, SX127x_BW_500_00_KHZ, SX127x_SF_6, SX127x_CR_4_7, 5000, TLM_RATIO_1_64, 4, 8},
     {1, RATE_100HZ, SX127x_BW_500_00_KHZ, SX127x_SF_7, SX127x_CR_4_7, 10000, TLM_RATIO_1_64, 4, 8},
@@ -84,23 +32,18 @@ static expresslrs_mod_settings_t air_rate_config[4] = {
     {3, RATE_25HZ, SX127x_BW_500_00_KHZ, SX127x_SF_9, SX127x_CR_4_7, 40000, TLM_RATIO_NO_TLM, 4, 10},
 };
 
-typedef struct {
-  int8_t index;
-  expresslrs_rf_rates_t rate; // Max value of 16 since only 4 bits have been assigned in the sync package.
-  int32_t rx_sensitivity;     //expected RF sensitivity based on
-  uint32_t toa;               //time on air in microseconds
-  uint32_t rf_mode_cycle_interval;
-  uint32_t rf_mode_cycle_addtional_time;
-  uint32_t sync_pkt_interval_disconnected;
-  uint32_t sync_pkt_interval_connected;
-} expresslrs_rf_pref_params_t;
-
 static const expresslrs_rf_pref_params_t rf_pref_params[4] = {
     {0, RATE_200HZ, -112, 4380, 3000, 2500, 2000, 4000},
     {1, RATE_100HZ, -117, 8770, 3500, 2500, 2000, 4000},
     {2, RATE_50HZ, -120, 17540, 4000, 2500, 2000, 4000},
     {3, RATE_25HZ, -123, 17540, 6000, 4000, 2000, 4000},
 };
+
+extern uint8_t fhss_index;
+extern void fhss_update_freq_correction(uint8_t value);
+extern void fhss_randomize(int32_t seed);
+extern uint32_t fhss_get_freq(uint16_t index);
+extern uint32_t fhss_next_freq();
 
 float rx_rssi;
 
@@ -111,21 +54,6 @@ static uint32_t current_rate = 0;
 static uint32_t max_sync_delay = 0;
 
 const uint8_t UID[6] = {223, 93, 241, 111, 2, 94};
-const uint32_t fhss_freqs[] = {
-    863275000, // band H1, 863 - 865MHz, 0.1% duty cycle or CSMA techniques, 25mW EIRP
-    863800000,
-    864325000,
-    864850000,
-    865375000, // Band H2, 865 - 868.6MHz, 1.0% dutycycle or CSMA, 25mW EIRP
-    865900000,
-    866425000,
-    866950000,
-    867475000,
-    868000000,
-    868525000, // Band H3, 868.7-869.2MHz, 0.1% dutycycle or CSMA, 25mW EIRP
-    869050000,
-    869575000,
-};
 
 const unsigned char crc8tab[256] = {
     0x00, 0xD5, 0x7F, 0xAA, 0xFE, 0x2B, 0x81, 0x54, 0x29, 0xFC, 0x56, 0x83, 0xD7, 0x02, 0xA8, 0x7D,
@@ -153,119 +81,9 @@ static inline uint8_t elrs_calc_crc(uint8_t *data, uint8_t length) {
   return crc;
 }
 
-#define freq_correction_max 200000
-#define freq_correction_min -200000
-
-static int32_t freq_correction = 0;
-
-void fhss_update_freq_correction(uint8_t value) {
-  if (!value) {
-    if (freq_correction < freq_correction_max) {
-      freq_correction += 61; //min freq step is ~ 61hz
-    } else {
-      freq_correction = freq_correction_max;
-      freq_correction = 0; //reset because something went wrong
-    }
-  } else {
-    if (freq_correction > freq_correction_min) {
-      freq_correction -= 61; //min freq step is ~ 61hz
-    } else {
-      freq_correction = freq_correction_min;
-      freq_correction = 0; //reset because something went wrong
-    }
-  }
-}
-
-static uint8_t fhss_index = 0;
-static uint8_t fhss_sequence[NR_SEQUENCE_ENTRIES] = {0};
-
-static uint32_t fhss_rng_seed = 0;
-
-int32_t fhss_rng() {
-  uint32_t m = 2147483648;
-  int32_t a = 214013;
-  int32_t c = 2531011;
-  fhss_rng_seed = (a * fhss_rng_seed + c) % m;
-  return fhss_rng_seed >> 16;
-}
-
-uint32_t fhss_rng_max(uint32_t max) {
-  uint32_t x = fhss_rng();
-  return (x * max) / FHSS_RNG_MAX;
-}
-
-uint32_t fhss_get_freq(uint16_t index) {
-  return fhss_freqs[index] - freq_correction;
-}
-
-void fhss_reset_is_available(uint8_t *is_available) {
-  // channel 0 is the sync channel and is never considered available
-  is_available[0] = 0;
-
-  for (uint32_t i = 1; i < NR_FHSS_ENTRIES; i++)
-    is_available[i] = 1;
-}
-
-void fhss_randomize() {
-  fhss_rng_seed = ((int64_t)UID[2] << 24) + ((int64_t)UID[3] << 16) + ((int64_t)UID[4] << 8) + UID[5];
-
-  uint8_t is_available[NR_FHSS_ENTRIES];
-  fhss_reset_is_available(is_available);
-
-  // Fill the FHSSsequence with channel indices
-  // The 0 index is special - the 'sync' channel. The sync channel appears every
-  // syncInterval hops. The other channels are randomly distributed between the
-  // sync channels
-  const int32_t SYNC_INTERVAL = NR_FHSS_ENTRIES - 1;
-
-  // for each slot in the sequence table
-  for (int32_t i = 0, prev = 0, left = NR_FHSS_ENTRIES - 1; i < NR_SEQUENCE_ENTRIES; i++) {
-    if (i % SYNC_INTERVAL == 0) {
-      // assign sync channel 0
-      fhss_sequence[i] = 0;
-      prev = 0;
-      continue;
-    }
-
-    // pick one of the available channels. May need to loop to avoid repeats
-    uint32_t index = 0;
-    do {
-      int32_t c = fhss_rng_max(left); // returnc 0 < c <left
-      // find the c'th entry in the isAvailable array
-      // can skip 0 as that's the sync channel and is never available for normal allocation
-      index = 1;
-
-      int32_t found = 0;
-      while (index < NR_FHSS_ENTRIES) {
-        if (is_available[index]) {
-          if (found == c)
-            break;
-          found++;
-        }
-        index++;
-      }
-      if (index == NR_FHSS_ENTRIES) {
-        index = 0;
-        break;
-      }
-    } while (index == prev); // can't use index if it repeats the previous value
-
-    fhss_sequence[i] = index; // assign the value to the sequence array
-    is_available[index] = 0;  // clear the flag
-    prev = index;             // remember for next iteration
-    left--;                   // reduce the count of available channels
-    if (left == 0) {
-      // we've assigned all of the channels, so reset for next cycle
-      fhss_reset_is_available(is_available);
-      left = NR_FHSS_ENTRIES - 1;
-    }
-  }
-}
-
-void elrs_set_frequency(uint32_t freq) {
+void elrs_set_frequency(int32_t FRQ) {
   sx12xx_set_mode(SX127x_OPMODE_STANDBY);
 
-  int32_t FRQ = ((uint32_t)((double)freq / (double)FREQ_STEP));
   uint8_t buf[4] = {
       (uint8_t)((FRQ >> 16) & 0xFF),
       (uint8_t)((FRQ >> 8) & 0xFF),
@@ -345,8 +163,7 @@ void elrs_hop() {
     return;
   }
 
-  fhss_index++;
-  elrs_set_frequency(fhss_get_freq(fhss_sequence[fhss_index]));
+  elrs_set_frequency(fhss_next_freq());
 
   if (air_rate_config[current_rate].tlm_interval == 0) {
     elrs_enter_rx();
@@ -444,7 +261,7 @@ void rx_init() {
     return;
   }
 
-  fhss_randomize();
+  fhss_randomize(((int64_t)UID[2] << 24) + ((int64_t)UID[3] << 16) + ((int64_t)UID[4] << 8) + UID[5]);
 
   elrs_set_rate(0);
   elrs_set_frequency(fhss_get_freq(0));
